Updated to use latest core package
